services/michelangelo/pyproject.toml
[tool.poetry]
name = "another_service_example"
version = "0.1.0"
description = ""
authors = ["nirmansour <nirooma@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.15.0"
fastapi = "^0.68.1"
tortoise-orm = "^0.17.7"
aerich = "^0.5.8"
psycopg2-binary = "^2.9.1"
asyncpg = "^0.24.0"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
kombu = "^5.1.0"
gunicorn = "^20.1.0"
tortoise-cli = "^0.1.1"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
requests = "^2.26.0"
pytest-asyncio = "^0.15.1"
asynctest = "^0.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
